+++ Fuji/Source/FileSystem/MFFileSystemMemory.cpp
@@ -109,57 +109,57 @@ int MFFileMemory_Close(MFFile* fileHandle)
 	return 0;
 }
 
-int MFFileMemory_Read(MFFile* fileHandle, void *pBuffer, int64 bytes)
+size_t MFFileMemory_Read(MFFile* fileHandle, void *pBuffer, size_t bytes)
 {
 	MFCALLSTACK;
 
 	MFFileMemoryData *pMem = (MFFileMemoryData*)fileHandle->pFilesysData;
 
-	int64 bytesToCopy = fileHandle->length > -1 ? MFMin(bytes, fileHandle->length - fileHandle->offset) : bytes;
+	size_t bytesToCopy = fileHandle->length > -1 ? MFMin(bytes, (size_t)(fileHandle->length - fileHandle->offset)) : bytes;
 
-	MFCopyMemory(pBuffer, (char*)pMem->pMemoryPointer + fileHandle->offset, (uint32)bytesToCopy);
+	MFCopyMemory(pBuffer, (char*)pMem->pMemoryPointer + fileHandle->offset, bytesToCopy);
 	fileHandle->offset += bytesToCopy;
 
-	return (int)bytesToCopy;
+	return bytesToCopy;
 }
 
-int MFFileMemory_Write(MFFile* fileHandle, const void *pBuffer, int64 bytes)
+size_t MFFileMemory_Write(MFFile* fileHandle, const void *pBuffer, size_t bytes)
 {
 	MFCALLSTACK;
 
 	MFFileMemoryData *pMem = (MFFileMemoryData*)fileHandle->pFilesysData;
 
-	int64 bytesToCopy = pMem->allocated ? MFMin(bytes, (int64)pMem->allocated - fileHandle->offset) : bytes;
+	size_t bytesToCopy = pMem->allocated ? MFMin(bytes, pMem->allocated - (size_t)fileHandle->offset) : bytes;
 
-	MFCopyMemory((char*)pMem->pMemoryPointer + fileHandle->offset, pBuffer, (uint32)bytesToCopy);
+	MFCopyMemory((char*)pMem->pMemoryPointer + fileHandle->offset, pBuffer, bytesToCopy);
 
 	fileHandle->offset += bytesToCopy;
 	fileHandle->length = MFMax(fileHandle->offset, fileHandle->length);
 
-	return (int)bytesToCopy;
+	return bytesToCopy;
 }
 
-int MFFileMemory_Seek(MFFile* fileHandle, int64 bytes, MFFileSeek relativity)
+uint64 MFFileMemory_Seek(MFFile* fileHandle, int64 bytes, MFFileSeek relativity)
 {
 	MFCALLSTACK;
 
-	int64 newPos = 0;
+	uint64 newPos = 0;
 
 	switch(relativity)
 	{
 		case MFSeek_Begin:
-			newPos = MFMin(bytes, fileHandle->length);
+			newPos = MFMin((uint64)bytes, fileHandle->length);
 			break;
 		case MFSeek_End:
-			newPos = MFMax((int64)0, fileHandle->length - bytes);
+			newPos = MFMax(0ULL, fileHandle->length - bytes);
 			break;
 		case MFSeek_Current:
-			newPos = MFClamp((int64)0, fileHandle->offset + bytes, fileHandle->length);
+			newPos = MFClamp(0ULL, fileHandle->offset + bytes, fileHandle->length);
 			break;
 		default:
 			MFDebug_Assert(false, "Invalid 'relativity' for file seeking.");
 	}
 
 	fileHandle->offset = newPos;
-	return (int)newPos;
+	return newPos;
 }
