+++ Fuji/Source/FileSystem/MFFileSystemZipFile_Internal.h
@@ -15,15 +15,15 @@ MFFile* MFFileSystemZipFile_Open(MFMount *pMount, const char *pFilename, uint32 
 
 int MFFileZipFile_Open(MFFile *pFile, MFOpenData *pOpenData);
 int MFFileZipFile_Close(MFFile* fileHandle);
-int MFFileZipFile_Read(MFFile* fileHandle, void *pBuffer, int64 bytes);
-int MFFileZipFile_Write(MFFile* fileHandle, const void *pBuffer, int64 bytes);
-int MFFileZipFile_Seek(MFFile* fileHandle, int64 bytes, MFFileSeek relativity);
+size_t MFFileZipFile_Read(MFFile* fileHandle, void *pBuffer, size_t bytes);
+size_t MFFileZipFile_Write(MFFile* fileHandle, const void *pBuffer, size_t bytes);
+uint64 MFFileZipFile_Seek(MFFile* fileHandle, int64 bytes, MFFileSeek relativity);
 bool MFFileZipFile_FindFirst(MFFind *pFind, const char *pSearchPattern, MFFindData *pFindData);
 bool MFFileZipFile_FindNext(MFFind *pFind, MFFindData *pFindData);
 void MFFileZipFile_FindClose(MFFind *pFind);
 
 // this is just for convenience sake, and not part of the main filesystem interface
-uint32 MFFileZipFile_GetSize(const char* pFilename);
+uint64 MFFileZipFile_GetSize(const char* pFilename);
 bool MFFileZipFile_Exists(const char* pFilename);
 
 #endif
